+++ CHANGELOG.md
@@ -2,6 +2,6 @@
 - Added oblivion grass blocks
 - Added oblivion stone blocks
 ### Fixed:
-- 
+- Fixed quicksand generation
 ### Changed:
 -